chore(idpe-17265): code cleanup to make the test matrix more explicit
